+++ tests/unittests/utilities/test_plot.py
@@ -123,6 +123,7 @@ from torchmetrics.image import (
     TotalVariation,
     UniversalImageQualityIndex,
 )
+from torchmetrics.multimodal import LipVertexError
 from torchmetrics.nominal import CramersV, FleissKappa, PearsonsContingencyCoefficient, TheilsU, TschuprowsT
 from torchmetrics.regression import (
     ConcordanceCorrCoef,
@@ -665,6 +666,12 @@ _text_input_4 = lambda: [["there is a cat on the mat", "a cat is on the mat"]]
             lambda: torch.randn(1, 100, 3),
             lambda: torch.randn(1, 100, 3),
             id="procrustes disparity",
+        ),
+        pytest.param(
+            partial(LipVertexError, mouth_map=[0, 1, 2, 3, 4]),
+            lambda: torch.randn(10, 100, 3),
+            lambda: torch.randn(10, 100, 3),
+            id="lip vertex error",
         ),
     ],
 )